Changes to the interface/design of the remappers, plus adding some default ones
It is now possible to set up the remapper in a gradual way. For instance,
you no longer_need_ to know in advance how many fields you will set in the
remapper. The remapper can dynamically "push_back" fields. The exact count
will be computed during the `registration_complete` call. Furthermore, you
no longer need to have the actual field for the registration phase. You can
register the field just with its FieldIdentifier, and bind the actual field
at a later stage (either before or after registration_complete, but, obviously,
*before* you try to call remap).

I also added implementation for identity, inverse, and combo remappers.
The first does close to nothing, and is used when src and tgt grids are
the same (actually, have the same name). The second simply wraps an existing
remapper, and uses it in the opposite direction. Finally, the third remapper
simply chains two user-provided remappers into a single operation (notice
that no optimization is done, the two remappers are simply called in sequence,
so no performance gain, just simplyfing the calls).