Adopt netcdf-c for all netcdf output.
This commit is the first step to not having any Fortran dependency
in the simple-io framework.  The next step will be to adopt netcdf-c
for the netcdf input.

This commit also adds netcdf output to the baseline testing for
p3_run_and_cmp.  But it does not yet actually write any data.  It
just creates the netcdf file where the data will be written.