Creates a simple IO interface for reading/writing netCDF files.
As a demonstration of the new interface the "p3_run_and_cmp"
test has been changed to use netCDF for writing baselines and
for later comparison against the baseline. The files produced
are also able to be compared using cprnc, however, automatic
use of cprnc has not been integrated in yet.

All new files are wrappers for the baseline netCDF-C commands to
make it easier to use and be more in line with the syntax and types
of variables we use in SCREAM.