SCREAM: Some changes in handling git stuff in the testing scripts
First, at the beginning of test-all, store original "ref" rather than
branch. If we're in a detached state (as it happens on jenkins),
the original branch does not exists locally (jenkins checks it out
in a detached state), so we may have issues checking it out later
(git may complain of a non-existing ref).

Second, when getting the current branch, do not rely on jenkins env
variables. If we're not on an actual local branch, then so be it,
we will use git refs instead. Using the env var GIT_BRANCH, makes
it look as if such branch locally existed (which it doesn't),
sparking issues when it's time to check it out (as stated above).