Add F2000 and rename FSCREAM compsets
Adds support for the old F2000 version of the SCREAM compsets (using the
existing use case file). We now have F2000 and F2010 versions of the
SCREAM compsets, so this required renaming the compset aliases to
distinguish between the two. So, we now have the following compsets:

    * F2000-SCREAM-LR
    * F2000-SCREAM-HR
    * F2000-SCREAM-LR-DYAMOND1
    * F2000-SCREAM-HR-DYAMOND1
    * F2010-SCREAM-LR
    * F2010-SCREAM-HR
    * F2010-SCREAM-LR-DYAMOND1
    * F2010-SCREAM-HR-DYAMOND1

The F2000 versions may go away at some point, but at least for the time
being this allows us to compare with the old compsets. The new naming
convention also makes it more clear what forcing is being used.