SCREAM: added SimpleGrid, upgraded PhysicsOnlyGridManager, and use them extensively
The simple grid does what the DummyPhysicsGrid used to do in several testing
units: a simple grid consisting of a bunch of columns. It is mostly meant
for testing individual parametrizations, since it offers all the interfaces
that the AD needs, without the need to setup and actual geometry.

Also, in several tests, we use the PhysicsOnlyGridManager instead of the
UserProvidedGridsManager, which is probably going to disappear soon, unless
we find a new purpose for it.