Clean up scorpio class, especially the init step.
Condensed the init step down to a set of internal functions.

Introduced m_dims map to class to handle the set of dimensions required
by each output stream.  As opposed to a paradigm where all possible dimensions
are included.

Note, dimensions are distinct from variables.  In PIO they are used to
determine the size and dimensionality of each variable.  If, for example, we
wanted to also write to file the lat and lon they would have to be added as
variables with dimension NCOL.