Fix RRTMG interface to be consistent with SPA changes for RRTMGP
Fix RRTMG interface to be consistent with changes made to
`radiation_init` in RRTMGP for SPA optics. This adds an additional
argument to `radiation_init` to pass `pbuf`. The edits to RRTMGP also
added a new namelist variable (`do_spa_optics`) that are not handled in
RRTMG. To get around this, we only add the namelist variable to the
namelist if using RRTMGP for the radiation package.