Fixes to the TimeStamp operators.
The operators lacked any reference to the timestamp month which was causing
false positives in the overloaded operator functions.

For simplification, a simple monotonic function is introduced that converts
the timestamp YYYYMMDD+SS to a double precision value that can be used in
operator comparisons.

Also a minor fix to the += operator that didn't consider that m_dd starts
at zero, while the number of days in a month start the day at 1.