Add psuedo-code for how SPA will be implemented.
More changes to the time manager - in this case create a function
to return the Julian Day, this will be needed for the temporal interpolation,
as well as by other routines unrealted to SPA.  For example, to calculate
the zenith angle in radiation.

Create the infrastructure for unit tests of SPA subfunctions.

placeholder so I can use this directory, not a true commit

remove big files