Remove the dependence of the scorpio input and output on the grids manager.
In the I/O interface the grids_manager was only really needed to
determine the global degree's of freedom each rank is responsbile
for.  Which is in turn used with PIO to read/write input/output.

In reality, it should be possible to pass an arbitrary grid with this
information, which would in principle allow I/O to handle files that
are not on the simulation grid.

In particular, this is a necessary precursor to full implementation
of SPA in SCREAM.