Finish implementation of loading in the SPA data from file for all months.
The next step will be to apply the remap weights to the spa data to
update the forcing used per SPA timestep.

This commit also changes how the spa data files are handled,
because they are too big to reside in the repo itself.

WARNING: We need to make this handling of the variables more general before this PR can go in.  Currently it only works with LC machines.