Final touches on setting up a CPRNC based comparison for multi-rank tests.
This commit establishes a cmake script in scream that will run a cprnc comparison
between two files.

It then uses this script in all of the tests that run through the SCREAM-AD and checks
all output from runs with MPI_RANKS>2 against the output from a run with just 1 rank.