Remove temporary override of tensor HV setting for ne4
Remove temporary override of namelist defaults to set tensor
hyperviscosity for ne4. This override appears to have been put in 11
months ago, with the note that it was a temporary fix for an upstream
merge, but I think we ultimately want to use tensor HV, and this is one
area that makes EAM compsets non-BFB between the E3SM and SCREAM repos.

[non-BFB]