Fix SPA timestate structure and spa_main to address test fails
This commit addresses two fails that show up in our testing on mappy.
In particular, we observe a valgrind fail and a fail in spa_main_test
for the release build.

1. The valgrind fail is due to a conditional on uninitialized memory.
   In this commit we initialize the SPA TimeState structure by setting
   the month to an impossible month value of -1 at construction.  That
   way in spa_update_data when it checks what month it is, there is
   initialized memory to compare with.
2. The spa_main_test has been updated to use the approx_equal function
   that is used in the common physics functions tests when comparing two
   values that should be equal.  Now we check the two values are equal to
   within a specific tolerance.

Additionally, for code cleanup, this commit adopts the routine defined in
common physic functions tests that creates a host mirror and immediately
deep copies.  Which cleans up the code a bit.