A hacky branch that makes it possible to examine water mass after a process
This branch is meant to facilitate writing output after both shoc and p3 to
examine the water mass before and after each process.  It is a simple change
that shouldn't go into master but will be useful for debugging until we
develop a more sophisticated infrastructure for writing output following a
process.