Adjust P3 interface to accumulate precip fluxes over subcycles.
This commit adjusts the interface of P3 to accumulate the overall
precipitation flux when P3 is subcycled.  It also resets the accumulated
fluxes to 0.0 whenever a new atmosphere timestep has begun.

Note.  This particular commit is a bit hacky.  There will be a separate PR
coming down the line that will clean up the interface and then we can remove
the hacky changes that this commit includes.