Switch precip in p3 to a mass, have surface cpl reset it to 0.0
This commit changes the definition of precip_liq_surf and
precip_ice_surf in the P3 interface and in surface coupling to be
a mass with units kg.  As such, the conversion in the surface coupling
export is changed.

Additionally, once the precipitation is used by surface coupling to
export to the component coupler, the surface coupling reset precip
to 0.0 so that the next round of precip accumulation in P3 starts from
0.0.

Note, in order to be able to convert precipitation mass from P3 into a
mass flux this commit added dt information to the surface coupling export.

Similarly, in order to take the precip flux provided by P3 and convert
it to a total mass the dt information needed to be passed to the p3
post-process structure.  For the latter, the dt is set everytime the
p3 run_impl is called.  This will allow for variable timestepping, if
we ever decide to implement that in the code.