Switch precipitation property check to a lower bound check.
This commit switches the precipitation property checks in P3 to be
lower bound based, since it is difficult to set an upper bound for
total mass accumulated without knowing the timestep.

This commit also adds a comment to the initialization of surface coupling
export and sets the local precipitation variables to be const in post-processing.