Change the precipitation variable names used locally to be more descriptive.
This commit changes the names of the precipitation variables used in the P3
interface to be more clear.  Local precipitation provided by p3_main is labeled
by `_flux` to denote a flux and the accumulated variable of precipitation that
will be provided to the rest of the model uses `_mass` since it is a mass.