Switch the pressure coordinate in P3 to be dry instead of wet.
Because the mass-mixing-ratios in P3 are dry, the pressure coordinates
are inconsistent, which leads to a mass leak during sedimentation in P3.

This commit introduces the pdel_dry and pmid_dry variables which are passed
to p3_main which represent the conversion of the wet pressure coordinates to
a dry set.