Add a function to SPA to organize remap data from file.
This commit adds a function to the SPA functions that organizes chunks
of spa remap data into segements.  This will be helpful in reducting the
memory footprint of the SPA read remap data function.

The approach is fashioned off of what is currently done in the component
coupler in handling mapping data for inter-component coupling.