Update tests to take advantage of the new system for remap indices
This commit also adjusts how the target column is defined in the SPA
Horizontal Interpolator structure.  Before this commit the target column
pointed to an index for the target column, rather than the actual
target column value.  The idea at the time was to have target column
be the index on this rank for the global target column.

The new interface is cleaner, letting the target column be the actual
target global column indicated by the remap.  Instead, there is a map
saved to the horizontal interpolator that stores the mapping from
global column id to the local index for this rank.