Establish a more robust random test for the horizontal remapping.
This commit improves the testing capabilities of the horizontal
remapper but creating a randomized test that checks that the remapper
performs well when initialized from data stored in views and from
data stored on file.

In the course of creating the tests some bugs were encountered which
are fixed in this commit.

The old test, which relied on a baseline file, has been removed.  As
well as the accompanying files.