Convert SPA interface to use new horizontal remapper.
This commit adjusts the SPA function which is responsible for reading
data from file and conducting the horizontal remap to use the new
horizontal remap structure.

In order to get SPA to work this commit also adds the ability to
"apply_remap" to a packed value as well as a 2d view and 3d view.