Fix spa tests to reflect use of horizontal remap utility.
This commit fixes the spa tests, which now use the horizontal remap
utility rather than the internal SPA based remapping logic.  This commit
also removes the test that checked the internal remapping, as this is now
redundant with the general horizontal remap test.

Lastly, this commit cleans up a lot of the spa code that is no longer
needed/used.