Add the capability to produce regional output using a remap file.
This commit will apply the new coarsening remapper to the scorpio
output interface to give the user the ability to produce remapped
output on the fly.

In order to trigger regional output the output control YAML file
needs to have the "remap_file" parameter given.  The value should
be a string that points to the remap file to be used.

Additionally all of the fields that will be written as output are
provided using "Field Names" as follows:
Field Names:
 - var1
 - var2
 - etc

NOT:
Fields:
  GRID:
    Field Names:
      - var1
      - var2
      - etc.