Make it possible to have diagnostics in regional output.
This commit fixes an issue where it was not possible to specify
diagnostics as part of remapped output.

The internal `get_field` function in scorpio_output.cpp is updated
to allow the user to specify a field mangager to get the field on,
instead of defaulting to the field manager for the output instance.

In the case of remapped data the stored field manager corresponds
to the fields as defined on the target grid.  We set a `sim` field
manager that stores the shared pointer to the main field manager
of the simulation.  When calculating diagnostics we do this on the
simulation field manager, not the io_field manager.