Fix the "number of columns" declaration in all unit tests to match init file.
This commit fixes the `input.yaml` files for all of the standalone tests to
reflect the actual number of columns in the initial condition file they are
using.

Note! This makes the unit tests larger since we were running with 218 columns
and using an ne4np4 init condition file (866 columns).  We could choose to go
 the other direction and keep the 218 columns and reduce the problems to ne2np4
size.