Add internal field managers to srfc export to track prescribed from file
This commit adds two internal field managers to the surface exporting
interface as a first step to tracking exported variables that will
be prescribed by file.  The purpose is to take advantage of the
numerous tools that can be used with fields.

The AtmosphereInput class is easy to use when passed a field manager.
The previous commit added the function `field_weighted_average` which
can provide a weighted average of two fields.  Combined we should be
able to read new input from file and then caculate the weighted averages
for all prescribed fields in a straightforward way.