Fix issue where some restart configurations were failing.
This commit fixes an issue where restarts that involved history files
with frequency and frequency units that didn't match the simulation
frequency and units would go looking for a restart history file even
if it was needed.  For example if we restart a monthly run with an
output stream that took average output every 3 hours then no restart
history file is needed but EAMxx would still go look for one in the
rpointer.atm file and throw an error.