Establish a time_interpolation class
This commit establishes a time_interpolation class which will be used
to store interpolation data for a set of variables and conduct a linear
time interpolation between data for those variables.

This first commit establishes a test for time interpolation and the begins
the actual class.  Currently, no data is read/stored, nor is the time interpolation
implemented.  The first step was to organize the data by timestamp to ensure that
when multiple files and timestamps are used the time interpolation is monotonic
with regards to the files and snaps it will read.

The sorting algorithm cycles through all of the files and determines the full list
of timestamps in the available data files and then sorts those timestamps, creating
a triplet structure for each timesnap with the filename, the timestamp and the time snap
in the respective file.  This can be used to progress through data steadily as the
simulation evolves.