Adjustments needed to apply interpolator in true EAMxx runs
This commit makes changes to the EAMxx surface exporter interface
to make it possible to apply the time interpolation of data from
file to the exported variables.

Note, this commit also keeps some code which is marked for deletion.
There will need to be a subsequent commit that cleans up the code.