Apply nudging timescale to the nudging process.
Before this commit the nudging process just did direct replacement, which
is not nudging per se.  With this commit the user is able to pass a
nudging timescale with runtime option `nudging_timescale` which will control
how strongly the nudging is applied.  Note, direct replacement is still an
option if the user chooses a nudging_timescale <= 0,