switch m_triplet_iterator in TimeInterpolation with a lookup index.
This commit removes the use of an iterator in the TimeInterpolation
class that tracks which set of data we are currently using.  The iterator
was causing issues in some use cases where the pointer was being
destroyed after the time interpolator was copied, but the original
was destroyed.  This commit gets around this issue.