Add an optional arg to surface coupling to support alternative field names.
This commit adds an optional argument to surface coupling that allows users
to specify an alternative name for variables that will be used when accessing
data from the source data files.  This is useful because typically source data
generated by a separate run won't have the variable names matching the
cryptic naming convention for surface export variables.  For example,
`Faxa_lwdn` vs. `lwdn`.

If the user leaves the new argument as NONE then the conventional names will be
used.