EAMxx: correctly handle data type in non-decomposed vars read/write
The user may be providing a data type that does not reflect what
is used in the NC file. While read/write_darray routines can handle
this, the non-decomposed get/put_var don't.
The user should not have to remember this detail, so we automatically
change the var dtype if needed, according to the provided input buffer.