Add runtime option to specify a file for static pressure in nudging.
This commit addresses a reviewer comment that the set of static pressure
levels may not always be stored with the nudging source data.  This commit
provides a runtime option to specify a specific file where the `p_levs`
information can be gathered.

This commit also refines the nudging pre-process for masked values to
always be consistent with the FillValue in the currently used file.