Add fill value support to time interpolation structure.
This commit adds support for time interpolation on data that is assigned
a filled value.  If interpolating between two values with 1 or more that
are "filled" than the resulting time interpolated value will be "filled"
as well.

This is necessary for dealing with time interpolation between source data
that could potentially be masked using a fill value.  If we don't take this
into consideration it can lead to crazy values in the output.