(INCOMPLETE) Add a unit test for vertical+horizontal remapping combined.
This commit is incomplete, there are still a few tasks that need to be
developed before this branch is ready.  This first step introduces a test
that conducts vertical and horizontal remapping combined.  Starting with
vertical remapping and then conducting a horizontal remap.  This tests
the configuration used by IO and in particular checks that values that
are masked in vertical remapping are treated appropriately in the
horizontal remap.