Response to review: Changed how defaults are handled for new params.
This commit removes the setting of the default shoc runtime options from
the initialization of SHOCRuntime structure and sets the defaults instead
in the namelist file itself.  This should make it easier for users to
see what the default values are at runtime.

Also a few minor fixes to remove left-over artifacts from earlier commits.