Increase resolution of rain fall speed lookup tables.
This commit increases the resolution and range of the rain fall speed
lookup tables, since the previous resolution caused significant
artifacts in the vertical profile of rain. The previous tables used
10 micron resolution for small drop sizes and 30 micron resolution for
larger drop sizes, covering a range between 5 and 8595 microns. The
new tables have a uniform 1 micron resolution between 1 and 9000
microns, which has proven sufficient to remove the numerical artifacts
in simple offline tests.

Since linear interpolation is used to extract fall speeds from the
lookup tables, and since the fall speed has negative second
derivative with respect to diameter over almost all of its range, this
commit systematically increases rain fall speed globally. However,
the climate impact of this effect is not expected to be large, since
it will generally only result in a small increase in rain evaporation.

This commit will require regeneration of P3 baselines, as well as
copying or regenerating the affected lookup tables to the input data
directory on all machines.

[CC]