EAMxx: add allow to call certain scorpio interfaces without pre-opening the file
When user simply wants to check a dim length, or see if the file has a dim/var,
it is tedious to register the file, call the function, then release the file.
Hence, we register the file and release it on the fly.
We can't register the file at the fcn entry point, since we don't know if the
file was already open, in which case we must register with the same mode.
So we create a "peek" helper local struct, that registers the file only if not open.
If the file had to be opened, when the struct goes out of scope (at return time),
the file will be released.