Ensure PIO calls are not optimized out
At least on some machines (mappy + gcc 9.2.0) if the return value
(ierr) from the PIO calls is ignored the PIO calls are optimized
out (values/attributes not written out to the file etc).

So adding a call to handle/touch the return value from the PIO
calls to ensure that values/attributes are updated in the output
file

Thanks lbertag@sandia.gov for the fix.